requirements.txt
requests>=2.25.1
pyyaml>=5.4.1
python-dotenv>=0.19.0